use session cookie to access security-enabled domain
