+++ kibana-reports/server/routes/utils/helpers.ts
@@ -20,9 +20,14 @@ import {
   ILegacyScopedClusterClient,
 } from '../../../../../src/core/server';
 import { RequestParams } from '@elastic/elasticsearch';
-import { CONFIG_INDEX_NAME, REPORT_STATE } from './constants';
+import {
+  CONFIG_INDEX_NAME,
+  REPORT_DEFINITION_STATUS,
+  REPORT_STATE,
+  TRIGGER_TYPE,
+} from './constants';
 import { CreateReportResultType } from './types';
-import { ReportSchemaType } from 'server/model';
+import { ReportDefinitionSchemaType, ReportSchemaType } from 'server/model';
 
 export function parseEsErrorResponse(error: any) {
   if (error.response) {
@@ -73,7 +78,7 @@ export const callCluster = async (
   return esResp;
 };
 
-export const saveToES = async (
+export const saveReport = async (
   isScheduledTask: boolean,
   report: ReportSchemaType,
   esClient: ILegacyClusterClient | ILegacyScopedClusterClient
@@ -98,7 +103,8 @@ export const saveToES = async (
   return esResp;
 };
 
-export const updateToES = async (
+// The only thing can be updated of a report instance is its "state"
+export const updateReportState = async (
   isScheduledTask: boolean,
   reportId: string,
   esClient: ILegacyClusterClient | ILegacyScopedClusterClient,
@@ -126,6 +132,65 @@ export const updateToES = async (
     updateParams,
     isScheduledTask
   );
+
+  return esResp;
+};
+
+export const saveReportDefinition = async (
+  reportDefinition: ReportDefinitionSchemaType,
+  esClient: ILegacyScopedClusterClient
+) => {
+  const toSave = {
+    report_definition: {
+      ...reportDefinition,
+      time_created: Date.now(),
+      last_updated: Date.now(),
+      status: REPORT_DEFINITION_STATUS.active,
+    },
+  };
+
+  const params: RequestParams.Index = {
+    index: CONFIG_INDEX_NAME.reportDefinition,
+    body: toSave,
+  };
+
+  const esResp = await esClient.callAsCurrentUser('index', params);
+
+  return esResp;
+};
+
+export const updateReportDefinition = async (
+  reportDefinitionId: string,
+  reportDefinition: ReportDefinitionSchemaType,
+  esClient: ILegacyScopedClusterClient
+) => {
+  let newStatus = REPORT_DEFINITION_STATUS.active;
+  /**
+   * "enabled = false" means de-scheduling a job.
+   * TODO: also need to remove any job in queue and release lock, consider do that
+   * within the createSchedule API exposed from reports-scheduler
+   */
+  if (reportDefinition.trigger.trigger_type == TRIGGER_TYPE.schedule) {
+    const enabled = reportDefinition.trigger.trigger_params.enabled;
+    newStatus = enabled
+      ? REPORT_DEFINITION_STATUS.active
+      : REPORT_DEFINITION_STATUS.disabled;
+  }
+
+  const toUpdate = {
+    report_definition: {
+      ...reportDefinition,
+      last_updated: Date.now(),
+      status: newStatus,
+    },
+  };
+
+  const params: RequestParams.Index = {
+    id: reportDefinitionId,
+    index: CONFIG_INDEX_NAME.reportDefinition,
+    body: toUpdate,
+  };
+  const esResp = await esClient.callAsCurrentUser('index', params);
 
   return esResp;
 };
